Use enum class for RCT2_SOUND

+++ src/openrct2/rct2/S6Exporter.cpp
@@ -918,9 +918,9 @@ void S6Exporter::ExportSpriteVehicle(RCT2SpriteVehicle* dst, const rct_vehicle* 
     dst->crash_x = src->crash_x;
     dst->sound2_flags = src->sound2_flags;
     dst->spin_sprite = src->spin_sprite;
-    dst->sound1_id = src->sound1_id;
+    dst->sound1_id = static_cast<uint8_t>(src->sound1_id);
     dst->sound1_volume = src->sound1_volume;
-    dst->sound2_id = src->sound2_id;
+    dst->sound2_id = static_cast<uint8_t>(src->sound2_id);
     dst->sound2_volume = src->sound2_volume;
     dst->sound_vector_factor = src->sound_vector_factor;
     dst->time_waiting = src->time_waiting;
@@ -930,7 +930,7 @@ void S6Exporter::ExportSpriteVehicle(RCT2SpriteVehicle* dst, const rct_vehicle* 
     dst->animation_frame = src->animation_frame;
     dst->var_C8 = src->var_C8;
     dst->var_CA = src->var_CA;
-    dst->scream_sound_id = src->scream_sound_id;
+    dst->scream_sound_id = static_cast<uint8_t>(src->scream_sound_id);
     dst->var_CD = src->var_CD;
     dst->var_CE = src->var_CE;
     dst->var_CF = src->var_CF;
